working auto update for windows

+++ src/Program.cs
@@ -120,11 +120,28 @@ static async Task Initialize(string currentVersion) {
 static async Task HandleArgs(string currentVersion) {
     string argName = "";
     if (Core.arguments.ContainsKey("--auto-update")) {
-        Cli.Output.Inform("Test 2");
-        Cli.Output.Inform($"Moving current file to {Core.arguments["--auto-update"].data!}");
-        File.Move(System.Diagnostics.Process.GetCurrentProcess().MainModule!.FileName!, Core.arguments["--auto-update"].data!);
-        Cli.Output.Inform("Successfully updated, start snipesharp as you would normally");
-        System.Diagnostics.Process.GetCurrentProcess().Kill();
+        try {
+            if (Cli.Core.pid == PlatformID.Unix) {
+                // make it work for unix
+            }
+            else {
+                // windows
+                // move file
+                File.Move(System.Diagnostics.Process.GetCurrentProcess().MainModule!.FileName!, Core.arguments["--auto-update"].data!);
+                FS.FileSystem.Log($"Moved {System.Diagnostics.Process.GetCurrentProcess().MainModule!.FileName!} to {Core.arguments["--auto-update"].data!}");
+
+                // output success
+                Cli.Output.Success("Successfully updated, start snipesharp again to run the latest version");
+
+                // kill current process
+                Console.ReadKey();
+                System.Diagnostics.Process.GetCurrentProcess().Kill();
+            }
+        }
+        catch (Exception e) { 
+            Cli.Output.Inform(e.ToString());
+            FS.FileSystem.Log(e.ToString());
+         }
     } 
     if (Core.arguments.ContainsKey("--username")) Config.v.DiscordWebhookUsername = Core.arguments["--username"].data!;
     if (Core.arguments.ContainsKey("--asc")) Config.v.AutoSkinChange = true;
